+++ examples/png/main.cpp
@@ -6,9 +6,9 @@
 #include <math.h>
 #include <complex>
 
-#include "gfxDraw.h"
-#include "gfxDrawCircle.h"
-#include "gfxDrawObject.h"
+#include "../../src/gfxDraw.h"
+#include "../../src/gfxDrawCircle.h"
+#include "../../src/gfxDrawObject.h"
 
 #include "lodepng.h"
 
@@ -41,8 +41,9 @@ void saveImage(char *fileName) {
   if (error) std::cout << "encoder error " << error << ": " << lodepng_error_text(error) << std::endl;
 }
 
-/// A lambda function to parse a parameter from the inputText.
-auto setImagePixel = [&](int16_t x, int16_t y, gfxDraw::RGBA color) {
+
+// Draw a pixel using a color.
+void setImagePixel(int16_t x, int16_t y, gfxDraw::RGBA color) {
   if ((x >= 0) && (x < imageWidth) && (y >= 0) && (y < imageHeight)) {
     image[4 * imageWidth * y + 4 * x + 0] = color.Red;
     image[4 * imageWidth * y + 4 * x + 1] = color.Green;
@@ -50,6 +51,17 @@ auto setImagePixel = [&](int16_t x, int16_t y, gfxDraw::RGBA color) {
     image[4 * imageWidth * y + 4 * x + 3] = 255;
   }
 };
+
+gfxDraw::RGBA readImagePixel(int16_t x, int16_t y) {
+  gfxDraw::RGBA col;
+  if ((x >= 0) && (x < imageWidth) && (y >= 0) && (y < imageHeight)) {
+    col.Red = image[4 * imageWidth * y + 4 * x + 0];
+    col.Green = image[4 * imageWidth * y + 4 * x + 1];
+    col.Blue = image[4 * imageWidth * y + 4 * x + 2];
+    col.Alpha = 255;
+  }
+  return (col);
+}
 
 void fillImage(gfxDraw::RGBA color) {
   for (unsigned y = 0; y < imageHeight; y++) {
@@ -59,12 +71,10 @@ void fillImage(gfxDraw::RGBA color) {
   }
 }
 
-#define bmpSet(col) [](int16_t x, int16_t y) { \
-  setImagePixel(x, y, (col)); \
-}
 
-#define bmpDraw() [](int16_t x, int16_t y, gfxDraw::RGBA col) { \
-  setImagePixel(x, y, col); \
+// helpful "inline" function to draw to the png image with a fixed color
+#define pngSetPixel(col) [](int16_t x, int16_t y) { \
+  setImagePixel(x, y, (col)); \
 }
 
 // ===== Test paths
@@ -91,9 +101,23 @@ const char *PiePath = R"==(
    M175,175 h-150 a150,150 0 1,0 150,-150 z
    )==";
 
+// A SVG path showing a sword
 const char *SwordPath = R"==(
 M 40 80 L 100 10 L 130 0 L 120 30 L 50 90 C 60 100 60 110 70 100 C 70 110 80 120 70 120 A 14 14 0 0 1 60 130 A 50 50 0 0 0 40 100 C 36 99 36 99 35 105 l -15 13 C 10 121 10 121 12 110 L 25 95 C 31 94 31 94 30 90 A 50 50 90 0 0 0 70 A 14 14 0 0 1 10 60 C 10 50 20 60 30 60 C 20 70 30 70 40 80 M 100 10 L 100 30 L 120 30 L 102 28 L 100 10z
 )==";
+
+// A SVG path defining the shape of a heard
+const char *heardPath = "M43 7 a1 1 0 00-36 36l36 36 36-36a1 1 0 00-36-36z";
+
+// A SVG path defining the shape of an arrow
+const char *arrowPath = "M0 0l12-12 0 8 22 0 0 8-22 0 0 8Z";
+
+const char *keyPath =
+  "M177 10l-87 128 1 1 16 11 88-128L177 10z"
+  "m45 31c-10 15-22 12-34 4l-24 36c11 8 18 18 9 31l21 15 14-21-7-5c-2 1-3 1-5 1-6 0-11-5-11-11 0-6 5-11 11-11s11 5 11 11c0 0 0 0 0 0l8 5 9-13-19-12 6-9 7 5 5-8 12 8 9-12-21-15z"
+  "M70 136c0 0 0 0 0 0 0 0-1 0-1 0-2 1-4 2-5 4-1 2-2 4-2 6 0 2 1 3 3 5l38 27c2 1 3 1 5 1 2-1 4-2 5-4 1-2 2-4 2-6 0-2 0-3-2-4L73 137c-1-1-2-1-3-1z"
+  "M41 158c-12 0-23 5-29 14-8 12-6 30 5 45-1 2-1 5-1 8 0 12 10 21 21 21 4 0 8-1 12-4 22 9 44 6 54-8 9-14 5-33-8-50l-32-22c-7-3-14-4-21-4 0 0-1 0-1 0z"
+  "m-4 55c7 0 12 5 12 12 0 7-5 12-12 12-7 0-12-5-12-12s5-12 12-12z";
 
 // =====
 
@@ -231,51 +255,51 @@ void drawTest01() {
 
   auto drawHelper = [](const char *pathText, int16_t x, int16_t y) {
     std::vector<gfxDraw::Segment> vSeg = gfxDraw::parsePath(pathText);
-    fillSegments(vSeg, x, y, bmpSet(gfxDraw::BLUE), bmpSet(gfxDraw::RGBA(255, 255, 128)));
+    fillSegments(vSeg, x, y, pngSetPixel(gfxDraw::RGBA_BLUE), pngSetPixel(gfxDraw::RGBA(255, 255, 128)));
   };
 
   TRACE("\nTest-01:\n");
   // 8*8 rectange needs 2 edges on horizontal lines
-  gfxDraw::drawSolidRect(10, 10, 10, 10, bmpSet(gfxDraw::SILVER));
+  gfxDraw::drawSolidRect(10, 10, 10, 10, pngSetPixel(gfxDraw::RGBA_SILVER));
   drawHelper("M1 1 h7 v7 h-7 z", 10, 10);
 
   TRACE("\nTest-02:\n");
   // 7*8 rectange needs 2 more points on horizontal lines
-  gfxDraw::drawSolidRect(20, 10, 10, 10, bmpSet(gfxDraw::GRAY));
+  gfxDraw::drawSolidRect(20, 10, 10, 10, pngSetPixel(gfxDraw::RGBA_GRAY));
   drawHelper("M1 1 h6 v7 h-6 z", 20, 10);
 
   TRACE("\nTest-03:\n");
   // closed W
-  gfxDraw::drawSolidRect(30, 10, 11, 10, bmpSet(gfxDraw::SILVER));
+  gfxDraw::drawSolidRect(30, 10, 11, 10, pngSetPixel(gfxDraw::RGBA_SILVER));
   drawHelper("M1 1 h8 v7 l-4,-4 l-4 4 z", 30, 10);
 
   TRACE("\nTest-04: closed M\n");
-  gfxDraw::drawSolidRect(41, 10, 11, 10, bmpSet(gfxDraw::GRAY));
+  gfxDraw::drawSolidRect(41, 10, 11, 10, pngSetPixel(gfxDraw::RGBA_GRAY));
   drawHelper("M1 1 l4,4 l4 -4 v7 h-8 z", 41, 10);
 
   TRACE("\nTest-05: route\n");
-  gfxDraw::drawSolidRect(52, 10, 11, 10, bmpSet(gfxDraw::SILVER));
+  gfxDraw::drawSolidRect(52, 10, 11, 10, pngSetPixel(gfxDraw::RGBA_SILVER));
   drawHelper("M5 1 l4,4 l-4 4 l-4 -4 z", 52, 10);
 
   TRACE("\nTest-06: |> symbol\n");
-  gfxDraw::drawSolidRect(63, 10, 10, 10, bmpSet(gfxDraw::GRAY));
+  gfxDraw::drawSolidRect(63, 10, 10, 10, pngSetPixel(gfxDraw::RGBA_GRAY));
   drawHelper("M1 1 l7,4 l-7 4 z", 63, 10);
 
   TRACE("\nTest-07: <| symbol\n");
-  gfxDraw::drawSolidRect(73, 10, 10, 10, bmpSet(gfxDraw::SILVER));
+  gfxDraw::drawSolidRect(73, 10, 10, 10, pngSetPixel(gfxDraw::RGBA_SILVER));
   drawHelper("M8 1 l-7,4 l7 4 z", 73, 10);
 
   TRACE("\nTest-10: [[]] \n");
-  gfxDraw::drawSolidRect(10, 22, 10, 10, bmpSet(gfxDraw::SILVER));
+  gfxDraw::drawSolidRect(10, 22, 10, 10, pngSetPixel(gfxDraw::RGBA_SILVER));
   drawHelper("M1 1 h7 v7 h-7 z M3 3 h3 v3 h-3 z", 10, 22);
 
   TRACE("\nTest-11: [*] \n");
-  gfxDraw::drawSolidRect(20, 22, 10, 10, bmpSet(gfxDraw::GRAY));
+  gfxDraw::drawSolidRect(20, 22, 10, 10, pngSetPixel(gfxDraw::RGBA_GRAY));
   drawHelper("M1 1 h7 v7 h-7 z M4 4 h1 v1 h-1 z", 20, 22);
 
-  gfxDraw::drawSolidRect(10, 34, 27, 24, bmpSet(gfxDraw::SILVER));
+  gfxDraw::drawSolidRect(10, 34, 27, 24, pngSetPixel(gfxDraw::RGBA_SILVER));
   gfxDraw::pathByText("M2 52 l60-50 v20 h20 l-30 50z", 10, 34, 30,
-                      bmpSet(gfxDraw::BLUE),
+                      pngSetPixel(gfxDraw::RGBA_BLUE),
                       [](int16_t x, int16_t y) {
                         gfxDraw::RGBA col = gfxDraw::RGBA(0xC0, 0x0, 0, 0xFF);
                         col.Green += (7 * x);
@@ -284,23 +308,23 @@ void drawTest01() {
                       });
 
 
-  gfxDraw::drawSolidRect(40, 34, 43, 26, bmpSet(gfxDraw::SILVER));
+  gfxDraw::drawSolidRect(40, 34, 43, 26, pngSetPixel(gfxDraw::RGBA_SILVER));
   gfxDraw::pathByText(
     "M2 42 l20-33 l20 38 l20 -38 l20 33 l0 -40 l-70 0 z",
     40, 34, 50,
-    bmpSet(gfxDraw::BLUE), bmpSet(gfxDraw::YELLOW));
+    pngSetPixel(gfxDraw::RGBA_BLUE), pngSetPixel(gfxDraw::RGBA_YELLOW));
 
   // radio with 2 inner paths
-  gfxDraw::drawSolidRect(91, 34, 43, 35, bmpSet(gfxDraw::SILVER));
-  gfxDraw::pathByText(RadioPath, 90, 34, 50, bmpSet(gfxDraw::BLACK), bmpSet(gfxDraw::YELLOW));
+  gfxDraw::drawSolidRect(91, 34, 43, 35, pngSetPixel(gfxDraw::RGBA_SILVER));
+  gfxDraw::pathByText(RadioPath, 90, 34, 50, pngSetPixel(gfxDraw::RGBA_BLACK), pngSetPixel(gfxDraw::RGBA_YELLOW));
 
-  // smiley with beziers
-  gfxDraw::drawSolidRect(10, 80, 99, 99, bmpSet(gfxDraw::GREEN));
-  gfxDraw::pathByText(SmilieCurvePath, 11, 81, 200, bmpSet(gfxDraw::BLUE), bmpSet(gfxDraw::YELLOW));
+  // smiley with bezier curves
+  gfxDraw::drawSolidRect(10, 80, 99, 99, pngSetPixel(gfxDraw::RGBA_GREEN));
+  gfxDraw::pathByText(SmilieCurvePath, 11, 81, 200, pngSetPixel(gfxDraw::RGBA_BLUE), pngSetPixel(gfxDraw::RGBA_YELLOW));
 
   // smiley with Arcs
-  gfxDraw::drawSolidRect(120, 80, 99, 99, bmpSet(gfxDraw::GREEN));
-  gfxDraw::pathByText(SmileyArcPath, 120, 81, 200, bmpSet(gfxDraw::BLUE), bmpSet(gfxDraw::YELLOW));
+  gfxDraw::drawSolidRect(120, 80, 99, 99, pngSetPixel(gfxDraw::RGBA_GREEN));
+  gfxDraw::pathByText(SmileyArcPath, 120, 81, 200, pngSetPixel(gfxDraw::RGBA_BLUE), pngSetPixel(gfxDraw::RGBA_YELLOW));
 
   // gfxDraw::drawCubicBezier(5, 10, 10, -5, 20, 25, 26, 10,
   //                          [](int16_t x, int16_t y) {
@@ -313,27 +337,27 @@ void drawTest01() {
 // draw the test paths for test02.png
 void drawTest02() {
   std::vector<gfxDraw::Segment> segs = gfxDraw::parsePath("M1 1 h7 v7 h-7 z M3 3 h3 v3 h-3 z");
-  gfxDraw::scaleSegments(segs, 400);                                                       // scale by 400%
-  gfxDraw::moveSegments(segs, 100, 0);                                                     // move to the right
-  gfxDraw::fillSegments(segs, bmpSet(gfxDraw::BLACK), bmpSet(gfxDraw::RGBA(0xEEEEEEFF)));  // hard-coded fill color here.
+  gfxDraw::scaleSegments(segs, 400);                                                                      // scale by 400%
+  gfxDraw::moveSegments(segs, 100, 0);                                                                    // move to the right
+  gfxDraw::fillSegments(segs, pngSetPixel(gfxDraw::RGBA_BLACK), pngSetPixel(gfxDraw::RGBA(0xEEEEEEFF)));  // hard-coded fill color here.
 
-  gfxDraw::moveSegments(segs, 22, 15);                                                     // move right down
-  gfxDraw::fillSegments(segs, bmpSet(gfxDraw::BLACK), bmpSet(gfxDraw::RGBA(0xEEEEEEFF)));  // hard-coded fill color here.
+  gfxDraw::moveSegments(segs, 22, 15);                                                                    // move right down
+  gfxDraw::fillSegments(segs, pngSetPixel(gfxDraw::RGBA_BLACK), pngSetPixel(gfxDraw::RGBA(0xEEEEEEFF)));  // hard-coded fill color here.
 
   segs = gfxDraw::parsePath("M1 1 h7 v7 h-7 z M3 3 h3 v3 h-3 z");
-  gfxDraw::scaleSegments(segs, 400);                                                       // scale by 400%
-  gfxDraw::rotateSegments(segs, 25);                                                       // rotate some degrees
-  gfxDraw::moveSegments(segs, 200, 40);                                                    // move right down
-  gfxDraw::fillSegments(segs, bmpSet(gfxDraw::BLACK), bmpSet(gfxDraw::RGBA(0xEEEEEEFF)));  // hard-coded fill color here.
+  gfxDraw::scaleSegments(segs, 400);                                                                      // scale by 400%
+  gfxDraw::rotateSegments(segs, 25);                                                                      // rotate some degrees
+  gfxDraw::moveSegments(segs, 200, 40);                                                                   // move right down
+  gfxDraw::fillSegments(segs, pngSetPixel(gfxDraw::RGBA_BLACK), pngSetPixel(gfxDraw::RGBA(0xEEEEEEFF)));  // hard-coded fill color here.
 
 
   segs = gfxDraw::parsePath(SmilieCurvePath);
   gfxDraw::scaleSegments(segs, 200);
-  gfxDraw::rotateSegments(segs, 25);                                           // rotate some degrees
-  gfxDraw::moveSegments(segs, 40, 40);                                         // move right down
-  gfxDraw::fillSegments(segs, bmpSet(gfxDraw::BLACK), bmpSet(gfxDraw::LIME));  // hard-coded fill color here.
+  gfxDraw::rotateSegments(segs, 25);                                                               // rotate some degrees
+  gfxDraw::moveSegments(segs, 40, 40);                                                             // move right down
+  gfxDraw::fillSegments(segs, pngSetPixel(gfxDraw::RGBA_BLACK), pngSetPixel(gfxDraw::RGBA_LIME));  // hard-coded fill color here.
 
-  gfxDraw::gfxDrawObject *rect = new gfxDraw::gfxDrawObject(gfxDraw::BLUE, gfxDraw::YELLOW);
+  gfxDraw::gfxDrawObject *rect = new gfxDraw::gfxDrawObject(gfxDraw::RGBA_BLUE, gfxDraw::RGBA_YELLOW);
   rect->setRect(32, 16);
 
   // rotate around center
@@ -343,30 +367,31 @@ void drawTest02() {
 
   rect->scale(200);
   rect->move(20, 20);
-  rect->draw(bmpDraw());
 
-  // // rect2->setFillGradient(gfxDraw::YELLOW, 10, 0, gfxDraw::ORANGE, 36, 0);
-  // rect2->setFillGradient(gfxDraw::YELLOW, 0, 2, gfxDraw::BLUE, 0, 22);
-  // rect2->draw(4, 240, bmpDraw());
+  rect->draw(setImagePixel);
 
-  // gfxDraw::rect(4, 200, 47, 31, nullptr, bmpSet(gfxDraw::SILVER));
-  // gfxDraw::pathByText("M0 12 l24-12 l20 20 h-16 v8 h-12 z", 5, 201, bmpSet(gfxDraw::BLUE), bmpSet(gfxDraw::YELLOW));
+  // // rect2->setFillGradient(gfxDraw::RGBA_YELLOW, 10, 0, gfxDraw::RGBA_ORANGE, 36, 0);
+  // rect2->setFillGradient(gfxDraw::RGBA_YELLOW, 0, 2, gfxDraw::RGBA_BLUE, 0, 22);
+  // rect2->draw(4, 240, setImagePixel);
+
+  // gfxDraw::rect(4, 200, 47, 31, nullptr, pngSetPixel(gfxDraw::SILVER));
+  // gfxDraw::pathByText("M0 12 l24-12 l20 20 h-16 v8 h-12 z", 5, 201, pngSetPixel(gfxDraw::RGBA_BLUE), pngSetPixel(gfxDraw::RGBA_YELLOW));
 
   // gfxDrawObject *rect1 = new gfxDrawObject(gfxDraw::TRANSPARENT, gfxDraw::SILVER);
   // rect1->setRect(46, 30);
-  // rect1->draw(4, 200, bmpDraw());
+  // rect1->draw(4, 200, setImagePixel);
 
-  // gfxDrawObject *draw1 = new gfxDrawObject(gfxDraw::BLUE, gfxDraw::ORANGE);
+  // gfxDrawObject *draw1 = new gfxDrawObject(gfxDraw::RGBA_BLUE, gfxDraw::RGBA_ORANGE);
   // draw1->setPath("M0 12 l24-12 l20 20 h-16 v8 h-12 z");
-  // draw1->draw(5, 201, bmpDraw());
+  // draw1->draw(5, 201, setImagePixel);
 
 
 #if 0
 
-  // gfxDraw::drawCubicBezier(80, 200, 80, 240, 150, 240, 150, 200, bmpSet(gfxDraw::BLACK));
+  // gfxDraw::drawCubicBezier(80, 200, 80, 240, 150, 240, 150, 200, pngSetPixel(gfxDraw::RGBA_BLACK));
 
-  // da->rect(180, 200, 84, 74, nullptr, bmpSet(gfxDraw::SILVER), 2);
-  // gfxDraw::pathByText("M80 200 C 80,240 150,240 150,200 z", 0,0, bmpSet(gfxDraw::BLACK), bmpSet(gfxDraw::YELLOW));
+  // da->rect(180, 200, 84, 74, nullptr, pngSetPixel(gfxDraw::RGBA_SILVER), 2);
+  // gfxDraw::pathByText("M80 200 C 80,240 150,240 150,200 z", 0,0, pngSetPixel(gfxDraw::RGBA_BLACK), pngSetPixel(gfxDraw::RGBA_YELLOW));
 #endif
 
 
@@ -375,61 +400,76 @@ void drawTest02() {
 };
 
 
-using namespace gfxDraw;
-
-// A SVG path defining the shape of a heard
-const char *heardPath = "M43 7 a1 1 0 00-36 36l36 36 36-36a1 1 0 00-36-36z";
-
-void drawTest04() {
-  // gfxDraw::pathByText(heardPath, 8, 8, 100, bmpSet(gfxDraw::BLUE), bmpSet(gfxDraw::YELLOW));
-
-  drawSolidRect(8, 8, 87, 80, bmpSet(SILVER));
-  // pathByText(heardPath, 8, 8, 100, bmpSet(BLUE), bmpSet(YELLOW));
-
-  // gfxDrawObject *widget = new gfxDrawObject();
-  // widget->setStrokeColor(YELLOW);
-  // widget->setFillColor(RED);
-  // widget->setPath(heardPath);
-  // widget->move(8, 8);
-  // widget->draw(bmpDraw());
-
-  gfxDrawObject widget;
-  widget.setStrokeColor(YELLOW);
-  widget.setFillColor(RED);
-  widget.setPath(heardPath);
-  widget.move(8, 8);
-  widget.draw(bmpDraw());
-}
-
 void drawTest03() {
 
   // gfxDraw::pathByText("M24 0c-14 0-24 10-24 24 c0 14 10 24 24 24 c14 0 24-10 24-24 c0-14-10-24-24-24Z",
-  // gfxDraw::pathByText("M24 0c-14 0-24 10-24 24  0 14 10 24 24 24  14 0 24-10 24-24  0-14-10-24-24-24Z", 11, 81, 200, bmpSet(gfxDraw::BLUE), bmpSet(gfxDraw::YELLOW));
+  // gfxDraw::pathByText("M24 0c-14 0-24 10-24 24  0 14 10 24 24 24  14 0 24-10 24-24  0-14-10-24-24-24Z", 11, 81, 200, pngSetPixel(gfxDraw::RGBA_BLUE), pngSetPixel(gfxDraw::RGBA_YELLOW));
 
   // Simple Arc drawing test cases
-  // gfxDraw::pathByText("M2 2 A 10 30 0 0 0 20 10 Z", 2, 2, 200, bmpSet(gfxDraw::BLUE), nullptr);
-  // gfxDraw::pathByText("M112 102 A 12 12 0 0 0 128 118 ", 2, 2, 100, bmpSet(gfxDraw::BLUE), nullptr);
+  // gfxDraw::pathByText("M2 2 A 10 30 0 0 0 20 10 Z", 2, 2, 200, pngSetPixel(gfxDraw::RGBA_BLUE), nullptr);
+  // gfxDraw::pathByText("M112 102 A 12 12 0 0 0 128 118 ", 2, 2, 100, pngSetPixel(gfxDraw::RGBA_BLUE), nullptr);
 
-  // gfxDraw::pathByText("M30 112 A 100 100 0 0 1 180 220 Z", 2, 2, 10, bmpSet(gfxDraw::BLUE), bmpSet(gfxDraw::YELLOW));
+  // gfxDraw::pathByText("M30 112 A 100 100 0 0 1 180 220 Z", 2, 2, 10, pngSetPixel(gfxDraw::RGBA_BLUE), pngSetPixel(gfxDraw::RGBA_YELLOW));
 
-  // gfxDraw::drawSolidRect(210, 10, 99, 99, bmpSet(gfxDraw::GREEN));
-  // gfxDraw::pathByText(SmilieCurvePath, 211, 11, 200, bmpSet(gfxDraw::BLUE), bmpSet(gfxDraw::YELLOW));
+  // gfxDraw::drawSolidRect(210, 10, 99, 99, pngSetPixel(gfxDraw::RGBA_GREEN));
+  // gfxDraw::pathByText(SmilieCurvePath, 211, 11, 200, pngSetPixel(gfxDraw::RGBA_BLUE), pngSetPixel(gfxDraw::RGBA_YELLOW));
 
-  // gfxDraw::drawSolidRect(210, 120, 99, 99, bmpSet(gfxDraw::GREEN));
+  // gfxDraw::drawSolidRect(210, 120, 99, 99, pngSetPixel(gfxDraw::RGBA_GREEN));
 
-  gfxDraw::pathByText(PiePath, 8, 8, 30, bmpSet(gfxDraw::BLUE), bmpSet(gfxDraw::YELLOW));
+  gfxDraw::pathByText(PiePath, 8, 8, 30, pngSetPixel(gfxDraw::RGBA_BLUE), pngSetPixel(gfxDraw::RGBA_YELLOW));
 
-  gfxDraw::pathByText(SwordPath, 10, 160, 100, bmpSet(gfxDraw::BLUE), bmpSet(gfxDraw::YELLOW));
+  gfxDraw::pathByText(SwordPath, 10, 160, 100, pngSetPixel(gfxDraw::RGBA_BLUE), pngSetPixel(gfxDraw::RGBA_YELLOW));
 
-  gfxDraw::pathByText(SmileyArcPath, 200, 100, 100, bmpSet(gfxDraw::BLUE), bmpSet(gfxDraw::YELLOW));
-  gfxDraw::pathByText(SmileyArcPath, 140, 10, 200, bmpSet(gfxDraw::BLUE), bmpSet(gfxDraw::GREEN));
-  gfxDraw::pathByText(SmileyArcPath, 220, 10, 50, bmpSet(gfxDraw::BLUE), bmpSet(gfxDraw::YELLOW));
-  gfxDraw::pathByText(SmileyArcPath, 100, 80, 150, bmpSet(gfxDraw::BLUE), bmpSet(gfxDraw::YELLOW));
+  gfxDraw::pathByText(SmileyArcPath, 200, 100, 100, pngSetPixel(gfxDraw::RGBA_BLUE), pngSetPixel(gfxDraw::RGBA_YELLOW));
+  gfxDraw::pathByText(SmileyArcPath, 140, 10, 200, pngSetPixel(gfxDraw::RGBA_BLUE), pngSetPixel(gfxDraw::RGBA_GREEN));
+  gfxDraw::pathByText(SmileyArcPath, 220, 10, 50, pngSetPixel(gfxDraw::RGBA_BLUE), pngSetPixel(gfxDraw::RGBA_YELLOW));
+  gfxDraw::pathByText(SmileyArcPath, 100, 80, 150, pngSetPixel(gfxDraw::RGBA_BLUE), pngSetPixel(gfxDraw::RGBA_YELLOW));
 
-  // gfxDraw::pathByText("O 20 20 20", 10, 120, 100, bmpSet(gfxDraw::BLUE), bmpSet(gfxDraw::YELLOW));
-  gfxDraw::pathByText("O 20 20 20 O 20 20 10", 10, 120, 100, bmpSet(gfxDraw::BLUE), bmpSet(gfxDraw::RGBA(0xFF, 0xFF, 0)));
+  // gfxDraw::pathByText("O 20 20 20", 10, 120, 100, pngSetPixel(gfxDraw::RGBA_BLUE), pngSetPixel(gfxDraw::RGBA_YELLOW));
+  gfxDraw::pathByText("O 20 20 20 O 20 20 10", 10, 120, 100, pngSetPixel(gfxDraw::RGBA_BLUE), pngSetPixel(gfxDraw::RGBA(0xFF, 0xFF, 0)));
 }
 
+
+using namespace gfxDraw;
+
+// draw a heard by using the gfxDrawObject functionality
+// demonstrate on how to undraw
+void drawTest04() {
+  // gfxDraw::pathByText(heardPath, 8, 8, 100, pngSetPixel(gfxDraw::RGBA_BLUE), pngSetPixel(gfxDraw::RGBA_YELLOW));
+
+  // setup the widget for drawing a heard
+  gfxDrawObject widget;
+  widget.setStrokeColor(RGBA_BLACK);
+  widget.setFillColor(RGBA_RED);
+  widget.setPath(heardPath);
+
+  // draw a background
+  drawSolidRect(10, 10, 87, 80, pngSetPixel(RGBA_SILVER));
+
+  // simple drawing on the rectangle behind
+  widget.move(10, 10);
+  widget.draw(setImagePixel);
+
+  // draw a rotated heard on top
+  widget.rotate(40, 44 + 10, 40 + 10);  // rotate on current center
+  widget.draw(setImagePixel);
+
+  // with undraw
+  // draw a background
+  drawSolidRect(110, 10, 87, 80, pngSetPixel(RGBA_SILVER));
+
+  // simple drawing on the rectangle behind
+  // no saving of background
+  widget.resetTransformation();
+  widget.move(110, 10);
+  widget.draw(setImagePixel, readImagePixel);
+  widget.undraw(setImagePixel);
+
+  // // draw a rotated heard on top
+  widget.rotate(40, 44 + 110, 40 + 10);  // rotate on current center
+  widget.draw(setImagePixel, readImagePixel);
+  widget.undraw(setImagePixel);
+}
 
 #if 0
 void drawClock(uint16_t _cx, uint16_t _cy, uint16_t _radius) {
@@ -446,7 +486,7 @@ void drawClock(uint16_t _cx, uint16_t _cy, uint16_t _radius) {
   da->drawCircle(
     _cx, _cy, _radius,
     [bmp](int16_t x, int16_t y) {
-      bmp->setPixel(x, y, gfxDraw::BLACK);
+      bmp->setPixel(x, y, gfxDraw::RGBA_BLACK);
     });
 
 
@@ -462,17 +502,17 @@ void drawClock(uint16_t _cx, uint16_t _cy, uint16_t _radius) {
     if ((i % 15) == 0) {
       // x1 = (x0 * 9.0) / 12.0;
       // y1 = (y0 * 9.0) / 12.0;
-      da->line(_cx + x0, _cy + y0, _cx + x1, _cy + y1, bmpSet(gfxDraw::BLUE), 5);
+      da->line(_cx + x0, _cy + y0, _cx + x1, _cy + y1, pngSetPixel(gfxDraw::RGBA_BLUE), 5);
 
     } else if ((i % 5) == 0) {
       // x1 = (x0 * 10.0) / 12.0;
       // y1 = (y0 * 10.0) / 12.0;
-      da->line(_cx + x0, _cy + y0, _cx + x1, _cy + y1, bmpSet(gfxDraw::GREEN), 3);
+      da->line(_cx + x0, _cy + y0, _cx + x1, _cy + y1, pngSetPixel(gfxDraw::RGBA_GREEN), 3);
 
     } else {
       // x1 = (x0 * 11.5) / 12.0;
       // y1 = (y0 * 11.5) / 12.0;
-      da->line(_cx + x0, _cy + y0, _cx + x1, _cy + y1, bmpSet(gfxDraw::BLACK), 1);
+      da->line(_cx + x0, _cy + y0, _cx + x1, _cy + y1, pngSetPixel(gfxDraw::RGBA_BLACK), 1);
     }
   }  // for
 }
@@ -481,18 +521,18 @@ void drawClock(uint16_t _cx, uint16_t _cy, uint16_t _radius) {
 
 
 int main() {
-  // silentTests();
+  silentTests();
 
-#if (1)
+#if (0)
   newImage(400, 300);
-  fillImage(gfxDraw::WHITE);
+  fillImage(gfxDraw::RGBA_WHITE);
   drawTest01();
   saveImage("test01.png");
 #endif
 
-#if (1)
+#if (0)
   newImage(400, 300);
-  fillImage(gfxDraw::WHITE);
+  fillImage(gfxDraw::RGBA_WHITE);
 
   drawTest02();
 
@@ -500,9 +540,9 @@ int main() {
 #endif
 
 
-#if (1)
+#if (0)
   newImage(400, 300);
-  fillImage(gfxDraw::WHITE);
+  fillImage(gfxDraw::RGBA_WHITE);
 
   drawTest03();
 
@@ -511,45 +551,45 @@ int main() {
 
 #if (1)
   newImage(400, 300);
-  fillImage(gfxDraw::WHITE);
+  fillImage(gfxDraw::RGBA_WHITE);
 
   drawTest04();
 
   saveImage("test04.png");
 #endif
 
-  // gfxDraw::rect(10, 50, 12, 8, bmpSet(gfxDraw::BLACK), bmpSet(gfxDraw::CYAN), 2);
-  // gfxDraw::rect(10, 60, 12, 8, nullptr, bmpSet(gfxDraw::CYAN), 2);
-  // gfxDraw::rect(10, 70, 12, 8, bmpSet(gfxDraw::BLACK), nullptr, 2);
+  // gfxDraw::rect(10, 50, 12, 8, pngSetPixel(gfxDraw::RGBA_BLACK), pngSetPixel(gfxDraw::CYAN), 2);
+  // gfxDraw::rect(10, 60, 12, 8, nullptr, pngSetPixel(gfxDraw::CYAN), 2);
+  // gfxDraw::rect(10, 70, 12, 8, pngSetPixel(gfxDraw::RGBA_BLACK), nullptr, 2);
 
-  // gfxDraw::rect(30, 70, 10, 8, bmpSet(gfxDraw::RED), nullptr, 2);
-  // gfxDraw::rect(39, 70, -10, 8, bmpSet(gfxDraw::BLACK), nullptr, 2);
+  // gfxDraw::rect(30, 70, 10, 8, pngSetPixel(gfxDraw::RED), nullptr, 2);
+  // gfxDraw::rect(39, 70, -10, 8, pngSetPixel(gfxDraw::RGBA_BLACK), nullptr, 2);
 
-  // gfxDraw::rect(42, 70, 10, 8, nullptr, bmpSet(gfxDraw::RED), 2);
-  // gfxDraw::rect(51, 70, -10, 8, nullptr, bmpSet(gfxDraw::BLACK), 2);
+  // gfxDraw::rect(42, 70, 10, 8, nullptr, pngSetPixel(gfxDraw::RED), 2);
+  // gfxDraw::rect(51, 70, -10, 8, nullptr, pngSetPixel(gfxDraw::RGBA_BLACK), 2);
 
-  // gfxDraw::rect(54, 70, 10, 8, nullptr, bmpSet(gfxDraw::RED), 2);
-  // gfxDraw::rect(54, 77, 10, -8, nullptr, bmpSet(gfxDraw::BLACK), 2);
+  // gfxDraw::rect(54, 70, 10, 8, nullptr, pngSetPixel(gfxDraw::RED), 2);
+  // gfxDraw::rect(54, 77, 10, -8, nullptr, pngSetPixel(gfxDraw::RGBA_BLACK), 2);
 
 
   // drawClock(120, 205, 80);
 
 
-  // drawDigits14(1, 0, 10, bmpSet(gfxDraw::RED), bmpSet(gfxDraw::YELLOW));
-  // drawDigits14(2, 80, 10, bmpSet(gfxDraw::RED), bmpSet(gfxDraw::YELLOW));
-  // drawDigits14(5, 160, 10, bmpSet(gfxDraw::RED), bmpSet(gfxDraw::YELLOW));
-  // drawDigits14(9, 250, 10, bmpSet(gfxDraw::RED), bmpSet(gfxDraw::YELLOW));
+  // drawDigits14(1, 0, 10, pngSetPixel(gfxDraw::RED), pngSetPixel(gfxDraw::RGBA_YELLOW));
+  // drawDigits14(2, 80, 10, pngSetPixel(gfxDraw::RED), pngSetPixel(gfxDraw::RGBA_YELLOW));
+  // drawDigits14(5, 160, 10, pngSetPixel(gfxDraw::RED), pngSetPixel(gfxDraw::RGBA_YELLOW));
+  // drawDigits14(9, 250, 10, pngSetPixel(gfxDraw::RED), pngSetPixel(gfxDraw::RGBA_YELLOW));
 
-  // drawDigits7(0, 0, 130, bmpSet(gfxDraw::RED), bmpSet(gfxDraw::YELLOW));
-  // drawDigits7(1, 34, 130, bmpSet(gfxDraw::RED), bmpSet(gfxDraw::YELLOW));
-  // drawDigits7(2, 68, 130, bmpSet(gfxDraw::RED), bmpSet(gfxDraw::YELLOW));
-  // drawDigits7(3, 102, 130, bmpSet(gfxDraw::RED), bmpSet(gfxDraw::YELLOW));
-  // drawDigits7(4, 136, 130, bmpSet(gfxDraw::RED), bmpSet(gfxDraw::YELLOW));
-  // drawDigits7(5, 170, 130, bmpSet(gfxDraw::RED), bmpSet(gfxDraw::YELLOW));
-  // drawDigits7(6, 204, 130, bmpSet(gfxDraw::RED), bmpSet(gfxDraw::YELLOW));
-  // drawDigits7(7, 238, 130, bmpSet(gfxDraw::RED), bmpSet(gfxDraw::YELLOW));
-  // drawDigits7(8, 282, 130, bmpSet(gfxDraw::RED), bmpSet(gfxDraw::YELLOW));
-  // drawDigits7(9, 316, 130, bmpSet(gfxDraw::RED), bmpSet(gfxDraw::YELLOW));
+  // drawDigits7(0, 0, 130, pngSetPixel(gfxDraw::RED), pngSetPixel(gfxDraw::RGBA_YELLOW));
+  // drawDigits7(1, 34, 130, pngSetPixel(gfxDraw::RED), pngSetPixel(gfxDraw::RGBA_YELLOW));
+  // drawDigits7(2, 68, 130, pngSetPixel(gfxDraw::RED), pngSetPixel(gfxDraw::RGBA_YELLOW));
+  // drawDigits7(3, 102, 130, pngSetPixel(gfxDraw::RED), pngSetPixel(gfxDraw::RGBA_YELLOW));
+  // drawDigits7(4, 136, 130, pngSetPixel(gfxDraw::RED), pngSetPixel(gfxDraw::RGBA_YELLOW));
+  // drawDigits7(5, 170, 130, pngSetPixel(gfxDraw::RED), pngSetPixel(gfxDraw::RGBA_YELLOW));
+  // drawDigits7(6, 204, 130, pngSetPixel(gfxDraw::RED), pngSetPixel(gfxDraw::RGBA_YELLOW));
+  // drawDigits7(7, 238, 130, pngSetPixel(gfxDraw::RED), pngSetPixel(gfxDraw::RGBA_YELLOW));
+  // drawDigits7(8, 282, 130, pngSetPixel(gfxDraw::RED), pngSetPixel(gfxDraw::RGBA_YELLOW));
+  // drawDigits7(9, 316, 130, pngSetPixel(gfxDraw::RED), pngSetPixel(gfxDraw::RGBA_YELLOW));
 
   printf("end.");
 }