+++ src/gfxDrawCommon.cpp
@@ -1,11 +1,11 @@
 // - - - - -
 // GFXDraw - A Arduino library for drawing shapes on a GFX display using paths describing the borders.
-// gfxDrawBezier.cpp: Library implementation file for bezier curve drawing functions
+// gfxDrawCommon.cpp: Library implementation for common functionality.
 //
 // Copyright (c) 2024-2024 by Matthias Hertel, http://www.mathertel.de
 // This work is licensed under a BSD style license. See http://www.mathertel.de/License.aspx
 //
-// Changelog: See gfxdraw.h and documentation files in this library.
+// Changelog: See gfxDraw.h and documentation files in this library.
 //
 // - - - - -