+++ src/gfxDrawObject.h
@@ -1,28 +1,21 @@
-// GFXDrawObject.h
-
-// This gfxDrawObject class implements the necessary steps to draw a object based on a path with transformation and
-// coloring options to simplify drawing with gfxDraw native functions.
-//
+// - - - - -
+// GFXDraw - A Arduino library for drawing shapes on a GFX display using paths describing the borders.
+// gfxDrawObject.cpp: Library implementation of a widget base class functionality
+// based on vector graphics and paths with transformation, coloring options to simplify drawing with gfxDraw functions.
 //  * All transformations are combined into a transformation matrix to avoid intermediate transformations with rounding
 //    effects.
 //  * The fill color can be specified using simple linear gradients.
 //  * The Transformations can be
 //  * Example of a gauge based on gfxDrawObject that manipulates segments based on a given value.
+//
+// Copyright (c) 2024-2024 by Matthias Hertel, http://www.mathertel.de
+// This work is licensed under a BSD style license. See http://www.mathertel.de/License.aspx
+//
+// Changelog: See gfxDraw.h and documentation files in this library.
 
 // https://svg-path-visualizer.netlify.app/#M2%2C2%20Q8%2C2%208%2C8
 
 #pragma once
-
-#ifdef ARDUINO
-#include <Arduino.h>
-#endif
-
-#include <functional>
-#include <vector>
-#include <algorithm>
-
-#include <cctype>
-// #include <cmath>
 
 #include "gfxDraw.h"
 #include "gfxDrawColors.h"
@@ -39,6 +32,10 @@ namespace gfxDraw {
 /// @brief Callback function definition to address a pixel on a display
 typedef std::function<void(int16_t x, int16_t y, RGBA color)> fDrawPixel;
 
+/// @brief Callback function definition to read a pixel from a display
+typedef std::function<RGBA(int16_t x, int16_t y)> fReadPixel;
+
+
 // Matrix type definition for transformation using 1000 factor and numbers.
 typedef int32_t Matrix1000[3][3];
 
@@ -47,7 +44,7 @@ typedef int32_t Matrix1000[3][3];
 /// It supports creating paths by simple functions like lines and rectangles or by using the svg path notation (not fully supported)
 /// The stroke color can be set.
 /// The fill color can be set to solid or gradient functions.
-/// TODO: collect the current colors from the canvas so the object can be "undrawn" later.
+/// TODO: collect the current colors from the canvas so the object can be "un-drawn" later.
 
 class gfxDrawObject {
 public:
@@ -83,18 +80,13 @@ public:
 
   // ===== path creation functions =====
 
-  // save the path by converting to a segment vector.
-  void setPath(const char *path) {
-    _segments = gfxDraw::parsePath(path);
-  }
+  /// @brief Create segments from a textual path definition.
+  void setPath(const char *path);
 
-  void setRect(int16_t w, int16_t h) {
-    _segments.push_back(gfxDraw::Segment(gfxDraw::Segment::Move, 0, 0));
-    _segments.push_back(gfxDraw::Segment(gfxDraw::Segment::Line, 0, h - 1));
-    _segments.push_back(gfxDraw::Segment(gfxDraw::Segment::Line, w - 1, h - 1));
-    _segments.push_back(gfxDraw::Segment(gfxDraw::Segment::Line, w - 1, 0));
-    _segments.push_back(gfxDraw::Segment(gfxDraw::Segment::Close));
-  }
+  /// @brief Create segments that form a with rectangle.
+  /// @param width width of the rectangle.
+  /// @param height height of the rectangle.
+  void setRect(int16_t width, int16_t height);
 
 
   // ===== coloring functions =====
@@ -176,7 +168,142 @@ public:
   };
 
 
-  void draw(gfxDraw::fDrawPixel cbDraw) {
+  class Background {
+  public:
+    // initialize without data.
+    Background() {
+      _w = 0;
+    };
+
+    void set(int16_t x, int16_t y, RGBA color) {
+      TRACE("bg::set(%d,%d)=%08x\n", x, y, color.raw);
+      if (_w == 0) {
+        _createData(x, y);
+        TRACE(" init to (%d/%d)-(%d/%d)\n", _x, _y, _x + _w - 1, _y + _h - 1);
+      }
+
+      if (x < _x) {
+        _insertLeft(_x - x);
+        TRACE(" resize to (%d/%d)-(%d/%d)\n", _x, _y, _x + _w - 1, _y + _h - 1);
+
+      } else if (x >= _x + _w) {
+        _insertRight(x - (_x + _w) + 1);
+        TRACE(" resize to (%d/%d)-(%d/%d)\n", _x, _y, _x + _w - 1, _y + _h - 1);
+      }
+
+      if (y < _y) {
+        // _insertTop(_y - y);
+        printf(" resize to %d/%d\n", x, y);
+        printf("undone!\n");
+
+      } else if (y >= _y + _h) {
+        _insertBottom(y - (_y + _h) + 1);
+        TRACE(" resize to (%d/%d)-(%d/%d)\n", _x, _y, _x + _w - 1, _y + _h - 1);
+      }
+
+      RGBA *cell = &data[(x - _x) + (y - _y) * _w];
+      if (cell->Alpha > 0) {
+        printf("double write at (%d,%d)!\n", x, y);
+      } else {
+        *cell = color;
+      }
+      // data[(x - _x) + (y - _y) * _w] = color;
+      // data.at((x - _x) + (y - _y) * _w) = color;
+    }
+
+
+    void draw(gfxDraw::fDrawPixel cbDraw) {
+
+      // insert blanks on each line
+      for (int16_t y = 0; y < _h; y++) {
+        for (int16_t x = 0; x < _w; x++) {
+          RGBA col = data[(y * _w) + x];
+          if (col.Alpha > 0)
+            cbDraw(x + _x, y + _y, col);
+        }
+      }
+
+    }  // draw();
+
+
+
+  private:
+    int16_t _x;
+    int16_t _y;
+    int16_t _w;
+    int16_t _h;
+    std::vector<RGBA> data;
+
+    // first initializing the data
+    void _createData(int16_t x, int16_t y) {
+      // first initialization.
+      _x = x & 0xFFF0;
+      _y = y;
+      _w = 16;
+      _h = 16;
+      data.resize(16 * 16);  // initialized with 0, all pixels with Alpha 0
+    };
+
+
+    // resize the array based 0/0
+    void _insertLeft(uint16_t count) {
+      RGBA blank('L', 'L', 'L', 0x00);
+
+      // expand by multiple of 16
+      count = (count + 15) & 0xFFF0;
+      int16_t x2 = _x - count;
+      int16_t w2 = _w + count;
+
+      TRACE(" l-expand %d -> %d\n", _w, w2);
+      data.reserve(w2 * _h);  // alloc at once.
+
+      // insert blanks on each line
+      for (int16_t l = 0; l < _h; l++) {
+        data.insert(data.begin() + (l * w2), count, blank);
+      }
+      _x = x2;
+      _w = w2;
+    }  // _insertLeft
+
+
+    // resize the array based 0/0
+    void _insertRight(uint16_t count) {
+      RGBA blank('R', 'R', 'R', 0x00);
+
+      // expand by multiple of 16
+      count = (count + 15) & 0xFFF0;
+      int16_t w2 = _w + count;
+      TRACE(" r-expand %d -> %d\n", _w, w2);
+
+      data.reserve(w2 * _h);  // alloc at once.
+
+      // insert blanks on each line
+      for (int16_t l = 0; l < _h; l++) {
+        data.insert(data.begin() + (l * w2) + _w, count, blank);
+      }
+
+      _w = w2;
+    }  // _insertRight
+
+
+    // resize the array based 0/0
+    void _insertBottom(uint16_t count) {
+      RGBA blank('b', 'b', 'b', 0x00);
+
+      // expand by multiple of 16
+      count = (count + 15) & 0xFFF0;
+      int16_t h2 = _h + count;
+      TRACE(" b-expand %d -> %d\n", _h, h2);
+
+      data.reserve(_w * h2);  // alloc at once.
+
+      data.insert(data.end(), count * _w, blank);
+      _h = h2;
+    }  // _insertBottom
+  };  // class Background
+
+
+  void draw(gfxDraw::fDrawPixel cbDraw, gfxDraw::fReadPixel cbRead = nullptr) {
     TRACE("draw()\n");
 
     TRACE(" stroke = %02x.%02x.%02x.%02x\n", _stroke.Alpha, _stroke.Red, _stroke.Green, _stroke.Blue);
@@ -195,31 +322,51 @@ public:
       y = ty / 1000;
     });
 
+    // create Background
+    if (cbRead) {
+      _bg = new Background();
+    }
+
     // draw...
     if (_fillColor1.Alpha == 0) {
       // need to draw the strike pixels only
       gfxDraw::drawSegments(tSegments, [&](int16_t x, int16_t y) {
+        if (cbRead) { _bg->set(x, y, cbRead(x, y)); }
         cbDraw(x, y, _stroke);
       });
 
     } else if ((_stroke.Alpha > 0)) {
       gfxDraw::fillSegments(
-        tSegments, [&](int16_t x, int16_t y) {
+        tSegments,
+        [&](int16_t x, int16_t y) {
+          if (cbRead) { _bg->set(x, y, cbRead(x, y)); }
           cbDraw(x, y, _stroke);
         },
         [&](int16_t x, int16_t y) {
+          if (cbRead) { _bg->set(x, y, cbRead(x, y)); }
           cbDraw(x, y, _getColor(x, y));
         });
 
     } else if (_stroke.Alpha == 0) {
       gfxDraw::fillSegments(
-        tSegments, nullptr,
+        tSegments,
+        nullptr,
         [&](int16_t x, int16_t y) {
+          if (cbRead) { _bg->set(x, y, cbRead(x, y)); }
           cbDraw(x, y, _getColor(x, y));
         });
     }
   };
 
+
+  void undraw(gfxDraw::fDrawPixel cbDraw) {
+    if (_bg) {
+      _bg->draw(cbDraw);
+    }
+    free(_bg);  //  = new Background();
+    _bg = nullptr;
+    printf("\n");
+  }
 
 
   void setFillGradient(gfxDraw::RGBA fill1, int16_t x1, int16_t y1, gfxDraw::RGBA fill2, int16_t x2, int16_t y2) {
@@ -377,7 +524,6 @@ private:
   gfxDraw::RGBA _stroke;
 
   // Fill coloring
-
   FillMode _fillMode;
 
   /// @brief used for solid filling and gradient start color.
@@ -395,6 +541,10 @@ private:
 
   // distance of the 2 gradient points.
   int32_t d1000;
+
+
+  // Reference to background data
+  class Background *_bg = nullptr;
 };
 
 // public:
@@ -428,7 +578,6 @@ private:
 //     drawCircleIntern(xc, yc, x, y, fBorderColor);
 //   }
 // }
-
 }
 
 // End.